// my_movie_quiz/frontend/myMovieQuiz/src/app/interfaces/actorTools.ts
export interface ActorTools {
    fontFamily: {index: number, 
                 value: string,
                 display: string,}
    background: { author_name: string;
                  highUrl: string;
                  id: number;
                  lowUrl: string;
                  stock_name: string;
                };
    backgrounds: any
    opacity: number;
    corner: { index: number;
              value: string;
            };
    weight: string;
    border: { index: number;
              value: string;
            };
    backColor: any;
    fontColor: any;
    borderColor: any;
    palette: string;
    colorArea: string;
  }